Modificacion del HandlerErrors

+++ src/services/product.Service.js
@@ -1,3 +1,5 @@
+const { CustomError } = require('../services/errors/CustomError');
+const { ErrorCodes } = require('../services/errors/errorCodes')
 
 class ProductService {
     constructor(storage) {
@@ -9,23 +11,29 @@ class ProductService {
     }
 
     getById(id) {
+        if (!id) {
+            throw new CustomError(ErrorCodes.INVALID_TYPES_ERROR, 'ID inválido');
+        }
         return this.storage.getById(id);
     }
 
     deleteById(id) {
+        if (!id) {
+            throw new CustomError(ErrorCodes.INVALID_TYPES_ERROR, 'ID inválido');
+        }
         return this.storage.deleteById(id);
     }
 
-    createOne(title, description, price, thumbnail, code, stock, category) {
+    createOne({ title, description, price, thumbnail, code, stock, category }) {
         if (!title || !description || !price || !thumbnail || !code || !stock || !category) {
-            throw new Error('invalid parameters');
+            throw new CustomError(ErrorCodes.MISSING_REQUIRED_FIELDS, 'Faltan campos requeridos');
         }
         return this.storage.createOne({ title, description, price, thumbnail, code, stock, category });
     }
 
     updateById(productId, updatedFields) {
         if (!productId || !updatedFields) {
-            throw new Error('invalid parameters');
+            throw new CustomError(ErrorCodes.INVALID_TYPES_ERROR, 'ID o campos de actualización inválidos');
         }
         return this.storage.updateById(productId, updatedFields);
     }
@@ -34,9 +42,6 @@ class ProductService {
         return this.storage.countDocuments(match);
     }
 
-    // find(match, { sort, skip, limit }) {
-    //     return this.storage.find(match, sort, skip, limit);
-    // }
 
     find(match, { sort, skip, limit }) {
         return this.storage.find(match)
